aws_autoscaling_group: deployed 2 ec2 with max 10
I had to diverge a bit from the book here, since the book has an older version
for creating the ec2 "templates", the book uses `aws_launch_configuration`
which has been deprecated or smth. I was getting 400s. Decided to update then
to aws_launch_template.

It seems that by doing this we do not need to specify the lifecycle part?
```
    # this bit, is specifically to handle how we update the ec2 when they are
    # being used by the ags.
    # when we make a change to this template and try to deploy it, aws wont be
    # able to destroy the ec2, because it is being used by the autoscaler.
    # Therefore we need to tell it to create the new ones and say to the ags to
    # use those new  ones, to then destroy the old ones.
    lifecycle {
        create_before_destroy = true
    }
```

So apparently it will do it by itself? maybe? need to look more into this.

Anyway, that is the only big change, all the rest keeps like the book. This
means the getting the subnets, is the same. Also the param name for the
security groups is different but other than that no big change.

Signed-off-by: bsantanad <benjamin.santana@proton.me>